errors: Support contextual error handling strategies
Currently `app_core::errors` handles synthesis of all error responses.
This means that this module needs to know about all possible error
types. To work around this, we wrap some errors with a special
`HttpError` type so this module can create a synthetic response.
Furthermore, this means that we can't handle errors differently in
inbound/outbound contexts.

This change modies the `error::respond` module with a new strategy
trait--`HttpRescue`--that is responsible for producing a
`SyntheticHttpResponse` if the error can be handled and returning an
error if it should not be handled gracefully. A `DefaultHttpRescue` type
encodes default behavior and inbound- and outbound-specific types extend
this functionality with appropriate error handling.

This change will enable us to perform error handling in more places in
the stack so that we can be more tactical about which errors are handled
in various parts of the stack--for example, so that error responses are
included in respose metrics.

This change modifies HTTP/1 error responses to include a `connection:
close` header when the server-side connection is being torn down.