Use a dedicated error type for HTTP response timeouts
We have several layers that are backed by `linkerd_timeout::Timeout` so
that its error type--ResponseTimeout--is somewhat ambiguous

This change consolidates the modules from `linkerd-timoeut` into
`linkerd-stack` to serve as a general timeout mechanism.
`linkerd_proxy_http::timeout` provides a stack that is parameterized
with a use-case-specific target and emits a use-case-specific error
type.

Error handlers are updated to use the HTTP-specific error type.