profiles: Share context-token allocations across instances
The profile client may be cloned to be shared across connections and
requests. This change stores the context token in an `Arc<str>` so
cloning gthe profile client doesn't require heap allocation. Now the
context token is re-allocated each time a gRPC request is issued. This
is much less frequent than the stack is cloned and also totally
unavoidable with prost.