inbound: Improve policy-based protocol detection
When an inbound server is configured with `proxyProtocol: HTTP/1`, we
have to do detection since client proxies may connect to us over HTTP/2.
But if the client is unmeshed or if detection fails, we can assume the
connection wasn't H2 upgraded.

This change modifies the inbound detection stack to use this improved
default behavior.