trace: remove `/tasks` admin endpoint and `tokio-trace` (#1265)
This branch removes the `/tasks` admin endpoint and the `tokio-trace`
crate that was used to implement it. The [`tokio-trace`] crate was a
rough prototype of what has now become the [Tokio Console] project, and
I'm now deprecating the experimental `tokio-trace` crate in favor of the
console. Also, the `tokio-trace` crate doesn't even _work_ anymore, as
the instrumentation in `tokio` that it used has been changed as part of
the Console work.

In 2.12, we'll replace this functionality by adding Tokio Console
support to the proxy. For now, though, we should remove the old stuff,
since it doesn't even work anymore.

This also means we can remove the special allow attribute for the git
dependency from the `cargo-deny` configuration. :)

[`tokio-trace`]: https://github.com/hawkw/tokio-trace
[Tokio Console]: https://github.com/tokio-rs/console