tracing: use `Span::or_current` when spawning tasks (#1272)
`tracing` v0.1.27 added a new [`Span::or_current`] method to aid with
efficiently propagating spans to spawned tasks (or threads). The method
is intended to be used when a spawned task should be instrumented with a
provided span if it is enabled, or the current span if the provided span
is _not_ enabled. `span.or_current()` returns `span` if it is enabled,
or, if `span` is disabled, it clones the current span and returns it.

This offers better performance than writing code like this:
```rust
    .instrument(span)
    .in_current_span()
```

First, `or_current` will only look up the current span (requiring a
thread-local access) _if_ `span` is not enabled, while
`.instrument(span).in_current_span()` will do this unconditionally.

Second, and more importantly, using `.in_current_span()` will wrap the
spawned future in _two_ `Instrumented` futures. One will enter the
current span whenever the task is polled, and inside that, the second
`instrument` will enter the passed span. This means we are entering two
spans instead of one, in the case where both are enabled. That adds a
small amount of additional overhead in every poll.

Therefore, this branch changes code that instruments a future prior to
spawning it to use `Span::or_current`. A lot of this is test code, which
(it turns out) doesn't really handle span propagation at all. However,
this also touches the main accept loop, as well as where we spawn admin
background tasks.

[`Span::or_current`]: https://docs.rs/tracing/0.1.27/tracing/span/struct.Span.html#method.or_current