Support a proxy-connection-close header
In 6cf1b2851, we updated the outbound proxy to automatically teardown
connections when the peer proxy sets the `l5d-proxy-error` response
header; but this header may be set in some situations when disconnecting
isn't appropriate and will only generate more load on the system.

This change introduces a `l5d-proxy-connection` header (like the
`connection` and `proxy-connection` headers) that the proxy can use to
signal that the proxied connection should be closed, independently of
the informational `l5d-proxy-error` message.

Some TODOs have been marked for followup improvements.