Avoid emitting informational headers to untrusted clients
The inbound proxy sets the `l5d-proxy-error` and `l5d-proxy-connection`
headers when an error is encountered. This leaks information about the
proxy to potentially untrusted clients.

This change modifies the error handling behavior to only emit these
headers from the inbound proxy when the client uses mesh identity.

In doing so, we modify the `inbound::NewSetIdentiyHeader` layer to use
`ExtractParam<tls::ConditionalServerTls, _>` to parameterize the server
connection's identity to be consistent with the rescue parameters. This
precipitated modifying the default impls for `ExtractParam` so that we
can provide an impl for `()` that simply uses the target type. A new
`CloneParam` type is introduced to replicate the former behavior.