outbound: Only honor the `l5d-proxy-connection` header when meshed (#1283)
The proxy uses the `l5d-proxy-connection` header to determine when
it should close proxied connections. This header must only be set by a
peer inbound proxy.

This change modifies the outbound proxy so that this connection is only
honored when the peer is actually meshed. If this header is set by an
unmeshed peer, the header is ignored.

This change also modifies the outbound proxy to always _clear_ this
header so that it can't be sent to the application.