Avoid attempting to reconnect to defunct endpoints (#1289)
Load balanced clients can get stuck continually trying to reconnect to
defunct endpoints in situations like the following:

1. A balancer has an existing endpoint but it is pending;
2. A new ready endpoint is added to the balancer;
3. Requests are sent to the new endpoint and the balancer ends up with
   the new endpoint being ready.
4. Service discovery issues an update indicating that we should no
   longer use the old endpoint, which has now shutdown.

This discovery update won't be processed by the balancer until we
attempt to issue a request on the balancer. But, because each endpoint
uses `SpawnReady` to attempt reconnection on a background task, we
continually attempt to reconnect to the defunct endpoint even though
we'll never actually issue requests to it (because it will be removed as
soon as the balancer is polled again).

There's a simple fix to this: we shouldn't put reconnection inside of
`SpawnReady`. Instead, we use a `SpawnReady` to drive each individual
connection attempt to readiness, but we don't drive reconnect to drive a
failed connection to be retried until the balancer has a chance to be
updated.

This may address linkerd/linkerd2#6842 and should fix another issue
reported in Slack where controller pods would continually log connection
failures after the destination controller was rescheduled.