tracing: use `Option`s as `tracing` values (#1299)
Tracing 0.1.29 added a `tracing::field::Value` implementation for
`Option<T> where T: Value`. This means that `Option`s can now be used as
structured field values, rather than recording them with `fmt::Debug`.
When the `Option` is `None`, the field will simply be skipped.

This allows us to make a bunch of existing code simpler. There are a
number of places where we create one span or event if some value is
`Some`, and a different span/event without that field if the value is
`None`. This can be changed to just use the `Option` as a field value
directly.

I'm almost certain I missed some of those places, but I changed the ones
I could find.

This also updates `tracing` to 0.1.29, of course.