Require identity configuration
The proxy currently supports a mode where identity is disabled. This
proliferates complexity that isn't really needed: there doesn't appear
to be a real use case where disabling identity is necessary. And, if it
is really necessary, we should reintroduce it after decoupling TLS and
identity.

This change causes the proxy to error during startup if identity is
disabled by configuration.

Furthermore, the `linkerd-proxy-identity` crate now has a `test-util`
feature that makes it possible to build a `LocalCrtKey` identity from
credentials provided by the `linkerd-identity/test-util` feature. A
default set of credentials are used in inbound and outbound tests.