dns-name: Remove `webpki` dependency
The `dns::Name` type is backed by `webpki::DNSName`; and the `webpki`
crate has a dependency on `ring`. As we setup to support alternate
cryptographic implementations, we don't want to incur this dependency
for such a simple type that only validates DNS-like names.

This change copies the `webpki::DNSName` and `webpki::DNSNameRef` types
as `dns::Name` and `dns::NameRef` (preserving copyright information).
Name parsing is simplified, as we don't need to handle wildcards with
these types.

Furthermore, this change updates the various identity-type wrapper types
to implement `Deref` so that `dns::Name::as_str` and
`dns::Name::as_bytes` are available implicitly.