Use jemalloc (#1321)
In load tests, it appears that jemalloc is--perhaps unsurprisingly--more efficient
than the alternatives. Specifically, jemalloc uses less memory and is more responsive
to memory being released without substantially impacting latency or CPU usage.

This is only enabled on `x86_64-unknown-linux-gnu`, as this is the only platform
on which all tests pass, etc (per [gnzlbg/jemallocator][readme]).

[readme]: https://github.com/gnzlbg/jemallocator/tree/c27a859e98e3cb790dc269773d9da71a1e918458#platform-support