Log discovery changes at DEBUG
It seems that we no longer log specific discovery changes at DEBUG
anywhere. We probably ought to revisit the discovery pipeline more
holistically and ensure that we're logging in strategic places, but for
now this will help users diagnose discovery updates without subscribing
to the much noisier TRACE stream.

Relates to linkerd/linkerd2#6842