Simplify retry policy
This change rewrites `RetryPolicy::can_retry` as
`RetryPolicy::body_too_big` and updates it to avoid reading
`content-length` header, instead relying on the `Body::size_hint` value
(which is set by the HTTP server if a `content-length` header is
present.

Renaming the function helps us express what the function actually checks
to help disambiguate it from all of the other 'retry' related
terminology in this file and is more explicit about what it does: we
don't precicesly know if the request is retryable, rather we know
whether the request is obviously too large to be retried. This also lets
us simplify comments and logging as the flow is more self-explanatory.

This change also adds unit tests for `RetryPolicy::body_too_big` to
exercise the variety of scenarios this function handles..

Furthermore, this change ensures that all `Body` implementations
correctly proxy `Body::size_hint`. `ReplayBody` is updated to cache the
the inner body's size hint value in the shared state so it can be reused
across all clones of the `ReplayBody` regardless of whether the current
clone holds the body state or not.