Use Body::size_hint to determine whether a request is retryable (#1345)
Currently, we look at each request's `content-length` header to
determine whether a request is too large to buffer for retries. This is
unnecessary, as hyper will already read the `content-length` and use it
populate the `Body::size_hint` value.

This change modifies the `ReplayBody` constructor (now
`ReplayBody::try_new`) to check the body's size hint and fails to
construct a `ReplayBody` if the lower bounds of the body size exceeds
the maximum buffer size.

In doing so, we update all `Body` implementations to ensure that
`size_hint` is properly implemenented. We also update the
`ReplayBody::size_hint` implementation to be more robust, returning the
original size hint if a clone of `ReplayBody` does not currently hold
the inner body.