`app-core` should not enable `meshtls-rustls` (#1360)
PR #1359 changed how `linkerd-meshtls`' TLS implementation feature flags
are enabled. After merging this PR, they should only be enabled by the
top level `linkerd2-proxy` crate, `linkerd-app-integration`, and any dev
dependencies that require a working TLS impl to be enabled for tests.
However, the `linkerd-app-core` crate already had a default-on feature
flag that enabled `meshtls-rustls`. I hadn't noticed this when working
on #1359, so I forgot to remove it.

This branch removes it.

Co-authored-by: Oliver Gould <ver@buoyant.io>