ci: Skip top-level crates in check-each-crate
The _check-each-crate_ job checks each crate individually to ensure its
dependencies are sufficient to compile the crate. This is less useful
for top-level crates, which are much slower to compile.

This change skips crates in `linkerd/app` and the top-level
`linkerd2-proxy` crate from individual checks.

These crates are exercised by the `check-clippy` job.