ci: Fix coverage workflow
The coverage workflow has been failing since boring was added as a
dependency.

This change updates the coverage workflow to include the proper build
dependencies and splits the build and test running phases into distinct
steps.

The tests are now run daily instead of weekly so we can more readily
detect and recover from regressions.