wip: Decouple profile router from caching strategy
The profile router eagerly instantiates all of its routes and reuses a
common inner service. This prevents routes from influencing which inner
service is used (i.e. by modifying the target being passed internally).

This change modifies the profile request router to create a new service
for each request if a route matches. A default route is always
instantiated (with no route value) so that this instantiation is only
needed when routes are being used.

This change temporarily disables route functionality.