ci: Skip unit tests in coverage report (#1390)
The coverage workflow is misconfigured so that it fails to run:
`cargo-tarpaulin` has command line arguments that don't match other
cargo commands (`--packages` instead of `--package`, which cannot be
mixed with `--workspace`). Furthermore, it doesn't seem that we actually
benefit from running tarpaulin over our non-integration tests, since the
coverage reports are not merged.