Handle opaque endpoints with known protocol
If an endpoint is set as opaque, but its logical service is not marked
as opaque, the connection will error out. The `TransportHeader` in this
case will not contain the logical name of the service, but it will still
do protocol detection.

Through this change, when an endpoint is marked as opaque and we connect
to the proxy's inbound port, if the `TransportHeader` has a protocol, we
go through the inbound HTTP stack instead of proxying TCP.
Add session protocol to local

Signed-off-by: Matei David <matei@buoyant.io>