Handle HTTP traffic over opaque transport connections (#1416)
If an endpoint is set as opaque, but its logical service is not marked
as opaque, clients may transport HTTP connections with an opaque
transport header. Inbound proxies currently fail to proxy these connections.

Through this change, when an inbound proxy receives a direct connection
with an opaque transport header--and that header includes a protocol and
no target service name--the connection will now be properly handled by the
HTTP stack.