Decouple client connection metadata from the I/O type
Some information about client connections may not be known a priori. For
example, the local socket address is only known after the connection is
established and the TLS handshake may negotiate additional information
(like the ALPN protocol).

We can use traits on the I/O type, but this means that all wrappers
types need implement all of these traits. This is cumbersome.

This change replaces the `tower::make::MakeConnection` helper trait with
a new `linkerd_stack::MakeConnection` trait that explicity returns a
`Metadata` instance so that connectors can return arbitrary information.

Now, the TCP connector returns a `Local<ClientAddr>` and the TLS
connector conditionally includes the negotiated protocol.

This change sets up the ability for the HTTP client to negotiate
protocol upgrading via ALPN instead of using per-request headers.