Use `AHash` for caches
[`AHash`][ahash] pruports to the be "the fastest DoS-resistant hash currently
available in Rust." It boasts substantial (~10x) latency improvements
over the default hash implementation (SipHash-1-3). It appears to be
an obvious choice for use in our maps, especially where hashing is
frequent and/or performed while a lock is held.

According to <https://lib.rs/crates/ahash>, it has substantial usage.

This change replaces the hash implementation for our metrics stores
and service cache (where locks are employed). It also updates the
service-profile router and traffic splitter, where potentially large
targets may be hashed for each request.

Other uses of the standard `HashMap`--mostly in metadata labels, where
hashing is not performed frequently--are currently left unchanged.

Signed-off-by: Oliver Gould <ver@buoyant.io>

[ahash]: https://github.com/tkaitchuck/aHash/blob/e77cab8c1e15bfc9f54dfd28bd8820c2a7bb27c4/compare/readme.md