access-log: Add JSON support (#1350)
This change adds the option to configure the access logging layer to
output JSON rather than the Apache Common access log format.
This is configured by the `LINKERD2_PROXY_ACCESS_LOG`
environment variable. When set to `json`, JSON-formatted logs are
emitted; and it must be set to `apache` for apache-formatted
logging.