build(deps): update tokio-util to v0.7 (#1490)
This branch updates `tokio-util` to v0.7.0.

The dependabot-generated PR for this update doesn't work, because this
is a breaking API change. The only update that was necessary here was
changing the use of `tokio_util::sync::PollSender` in
`linkerd-proxy-discover` to use the updated method names.

Closes #240

Signed-off-by: Eliza Weisman <eliza@buoyant.io>
(cherry picked from commit 7f173183b9941252901f67b44db151c9124fe146)
Signed-off-by: Oliver Gould <ver@buoyant.io>